% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TableIndicators.R
\name{TableIndicators}
\alias{TableIndicators}
\title{TableIndicators()
This function will return a table with most basic epi indicators for a specific country}
\usage{
TableIndicators(country)
}
\arguments{
\item{country}{the country to build the table for}
}
\description{
TableIndicators()
This function will return a table with most basic epi indicators for a specific country
}
